+++ tools/monitor/model/src/test/java/is/codion/tools/monitor/EntityServerMonitorTest.java
@@ -113,7 +113,7 @@ public class EntityServerMonitorTest {
 						.adminPort(3223)
 						.adminUser(User.parse("scott:tiger"))
 						.connectionPoolUsers(Collections.singletonList(UNIT_TEST_USER))
-						.domainClassNames(Collections.singletonList(TestDomain.class.getName()))
+						.domainClasses(Collections.singletonList(TestDomain.class.getName()))
 						.database(Database.instance())
 						.sslEnabled(false)
 						.build();