+++ tools/monitor/ui/src/test/java/is/codion/tools/monitor/ui/EntityServerMonitorPanelTest.java
@@ -92,7 +92,7 @@ public class EntityServerMonitorPanelTest {
 						.adminPort(3223)
 						.adminUser(User.parse("scott:tiger"))
 						.connectionPoolUsers(Collections.singletonList(UNIT_TEST_USER))
-						.domainClassNames(Collections.singletonList(TestDomain.class.getName()))
+						.domainClasses(Collections.singletonList(TestDomain.class.getName()))
 						.database(Database.instance())
 						.sslEnabled(false)
 						.build();